test: split combination and scenario tests
